Moved common tests to FieldHandlerTest (task #5061)
Before we had the common FieldHandlerTest, common field handler
tests were all in StringFieldHandlerTest or duplicated across
separate test field handler test classes.  Now they are all at
home in FieldHandlerTest class.
