+++ tests/TestCase/FieldHandlers/FieldHandlerTest.php
@@ -2,6 +2,7 @@
 namespace CsvMigrations\Test\TestCase\FieldHandlers;
 
 use CsvMigrations\FieldHandlers\Config\ConfigFactory;
+use CsvMigrations\FieldHandlers\Config\ConfigInterface;
 use CsvMigrations\FieldHandlers\FieldHandler;
 use CsvMigrations\FieldHandlers\FieldHandlerInterface;
 use PHPUnit_Framework_TestCase;
@@ -24,5 +25,20 @@ class FieldHandlerTest extends PHPUnit_Framework_TestCase
     {
         $this->assertFalse(empty($this->fh), "FieldHandler is empty");
         $this->assertTrue($this->fh instanceof FieldHandlerInterface, "FieldHandler does not implement FieldHandlerInterface");
+    }
+
+    public function testGetConfig()
+    {
+        $result = $this->fh->getConfig();
+
+        $this->assertInstanceOf(ConfigInterface::class, $result);
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     */
+    public function testRenderValueMissingRendererException()
+    {
+        $result = $this->fh->renderValue('test', ['renderAs' => 'thisRendererDoesNotExist']);
     }
 }
